+++ src/preview/styles.css
@@ -53,16 +53,3 @@ pre code {
 .fade-in {
   animation: fadeIn 300ms ease-in-out;
 }
-
-@keyframes slideIn {
-  from {
-    transform: translateX(-100%);
-  }
-  to {
-    transform: translateX(0%);
-  }
-}
-
-.slide-in {
-  animation: slideIn 0.5s ease-in-out;
-}
